led: Fix GetTrigger()
led.GetTrigger() considers that the `trigger` file contains a single
entry corresponding to the current trigger  for the given led.
This does not match my observations on 2 machines (x86_64 laptop and
aarch64 HA image booted using UEFI), the `trigger` file contains the
list of all valid trigger values, and the current trigger is between []

On the Green, this bug only impacts the initial state of the led state
in the dbus API: when the agent starts, it reads the `trigger` file, and
compares it against `none`. With the bug, even if the led initial
trigger is `none`, this will be returned as `... [none] ...` so the
led will mistakenly be considered as being on.

Then the rest of the code hardcodes the default value for each Green
led, and either writes this default value, or `none`, so it will work
fine.